requirements.txt
streamlit==1.18.1
protobuf==3.20.3
click==8.1.3
azure-ai-formrecognizer==3.3.0
azure-common==1.1.28
azure-core==1.33.0
python-dotenv==0.20.0
openai==0.27.0
altair==4.2.0
azure-core>=1.29.4
pdf2image==1.17.0
streamlit-extras
python-multipart
Pillow
pytesseract
